Round the dimensions of a newly created image variation
If the newly calculated dimension of an image is a floating point
number and by a fraction smaller than the specified/wanted dimension,
it will be taken into consideration and converted into an integer.

For instance, if the original image is a portrait image of 666x1000px,
and the desired variation is a landscape image of 300x180px, the
original image will be scaled to 666x399px before calculating the
variation. The newly calculated height of the variation is roughly
179.73, making it smaller than the specified height and therefore the
new height.

Since the difference between the calculated and the specified height
is infinitesimal, the proposed solution is to round the value.

A special thanks to:
* Stjepan Zlodi