Construct the directory path using the MD5 hash algorithm
Construct the directory path where the adjusted images will be saved to
using the MD5 hash algorithm.

Can be customized using the DAGUERRE_PATH setting set in the project's
settings. If left unspecified, the default value will be used, i.e
'dg'. The maximum length of the specified string is 13 characters.

Example:
* Default: dg/ce/2b/7014c0bdbedea0e4f4bf.jpeg
* Custom (DAGUERRE_PATH = 'img'): img/ce/2b/7014c0bdbedea0e4f4bf.jpeg

Known issue:
If the extracted hash string is 'ad', ad blockers will block the
image. All occurrences of 'ad' will be replaced with 'ag' since the MD5
hash produces letters from 'a' to 'f'.