Improve Unicode compatibility
Improve the models' Unicode compatibility and convert the source string
to bytes before hashing.

Due to Unicode-string-bytes changes introduced in Python 3, the hash
algorithm throws a TypeError when run on Python 3, complaining that
Unicode objects must be encoded before hashing. A quick view at the
Django's source code revealed that the accepted approach is to convert
the string to bytes.

References:
* Source code for django.contrib.auth.hashers